+++ models/view/src/plugin.ts
@@ -40,9 +40,6 @@ export default mergeIds(viewId, view, {
     Open: '' as ViewAction
   },
   component: {
-    ObjectFilter: '' as AnyComponent,
-    ValueFilter: '' as AnyComponent,
-    TimestampFilter: '' as AnyComponent,
     StringEditor: '' as AnyComponent,
     StringEditorPopup: '' as AnyComponent,
     StringPresenter: '' as AnyComponent,