+++ plugins/tracker-resources/src/viewOptions.ts
@@ -1,0 +1,7 @@
+import { ViewOptions } from '@anticrm/tracker'
+import { writable } from 'svelte/store'
+
+/**
+ * @public
+ */
+export const viewOptionsStore = writable<ViewOptions>()